refactor(desktop): split browser views from file browser container
